refactor(errors): move context to pkg
