+++ internal/store/backlog/errors.go
@@ -1,7 +1,7 @@
 package backlog
 
 import (
-	"nimona.io/internal/errors"
+	"nimona.io/pkg/errors"
 )
 
 var (
