+++ internal/store/graph/errors.go
@@ -1,7 +1,7 @@
 package graph
 
 import (
-	"nimona.io/internal/errors"
+	"nimona.io/pkg/errors"
 )
 
 const (
