+++ pkg/crypto/verify.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"math/big"
 
-	"nimona.io/internal/errors"
+	"nimona.io/pkg/errors"
 	"nimona.io/pkg/object"
 )
