+++ pkg/exchange/errors.go
@@ -1,7 +1,7 @@
 package exchange
 
 import (
-	"nimona.io/internal/errors"
+	"nimona.io/pkg/errors"
 )
 
 var (
