+++ pkg/exchange/exchange.go
@@ -11,7 +11,7 @@ import (
 	"github.com/gobwas/glob"
 
 	"nimona.io/pkg/context"
-	"nimona.io/internal/errors"
+	"nimona.io/pkg/errors"
 	"nimona.io/pkg/log"
 	"nimona.io/internal/store/graph"
 	"nimona.io/pkg/crypto"
