+++ pkg/exchange/manager.go
@@ -3,7 +3,7 @@ package exchange
 import (
 	"sync"
 
-	"nimona.io/internal/errors"
+	"nimona.io/pkg/errors"
 	"nimona.io/pkg/net"
 )
