+++ pkg/net/network.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"sync"
 
-	"nimona.io/internal/errors"
+	"nimona.io/pkg/errors"
 	"nimona.io/pkg/log"
 	"nimona.io/pkg/crypto"
 	"nimona.io/pkg/discovery"
