+++ pkg/orchestrator/queue.go
@@ -2,7 +2,7 @@ package orchestrator
 
 import (
 	"nimona.io/pkg/context"
-	"nimona.io/internal/errors"
+	"nimona.io/pkg/errors"
 )
 
 type (
